move page components to pages and dice roller to utils
